Split `UserTypesCodecTest`
